DispatchesTest and LoginTest moved to unittests
