+++ Engineering-61-Final-Project/src/test/java/com/sparta/eng61/unittests/DispatchesTest.java
@@ -1,4 +1,4 @@
-package com.sparta.eng61;
+package com.sparta.eng61.unittests;
 
 import com.sparta.eng61.pageobjects.dispatcher.DispatchesPage;
 import org.junit.jupiter.api.Assertions;
