+++ Engineering-61-Final-Project/src/test/java/com/sparta/eng61/unittests/LoginTest.java
@@ -1,4 +1,4 @@
-package com.sparta.eng61;
+package com.sparta.eng61.unittests;
 
 import com.sparta.eng61.pageobjects.dispatcher.LoginPage;
 import com.sparta.eng61.pageobjects.dispatcher.LogoutPage;
